Stop using member variables T, N & S in ext2 class (!119)
FileSystem member variables T, N & S are being used like local variables
in many of the file system specific set_used_sectors() methods.  They
are only used within each set_used_sectors() call and not used to
represent persistent information of a FileSystem interface class or to
pass information between separate methods.  Therefore stop using them
and replace them with local variables instead.

This block of code finds a field in the output and scans the number:
    Glib::ustring::size_type index = output.find("Block count:");
    if (index >= output.length() ||
        sscanf(output.substr(index).c_str(), "Block count: %lld", &T) != 1)
            T = -1;
The if statement says "if leading text is not found or scanning the
number fails then assign -1".  A sequence of two negatives leading to
assigning an error value is hard to understand.  Instead this an
equivalent block from btrfs::set_used_sectors():
    long long total_bytes = -1;
    Glib::ustring::size_type index = output.find("\ntotal_bytes");
    if (index < output.length())
            sscanf(output.substr(index).c_str(), "\ntotal_bytes %lld", &total_bytes);
This assigns a default error value and the if statement says "if leading
text found then scan the number".  Much simpler to understand.
Therefore change the code around to use this same pattern.

Anchor the leading text matches to the start of a new line in the
output where possible.  Just because it's what some of the other file
system's set_used_sectors() methods do (btrfs, reiser4 and xfs) and it
seems like more robust text matching.

Closes !119 -  Tidy-ups for file system interface classes