Pass structured device and partition names when loading FS_Info cache (#310)
Blkid is again falsely reporting ZFS on a whole disk device when a
partition contains a ZFS member [1].  See following commit "Implement
blkid false detection of whole disk ZFS workaround (#310)" for a full
explanation of the background and steps to re-create the issue.
    # blkid /dev/sdb*
    /dev/sdb: LABEL="testpool" UUID="5655469295808078251" UUID_SUB="9833733367633006209" BLOCK_SIZE="512" TYPE="zfs_member" PTUUID="63ca6054" PTTYPE="dos"
    /dev/sdb1: UUID="A829-6677" BLOCK_SIZE="512" TYPE="vfat" PARTUUID="63ca6054-01"
    /dev/sdb2: LABEL="testpool" UUID="5655469295808078251" UUID_SUB="9833733367633006209" BLOCK_SIZE="512" TYPE="zfs_member" PARTUUID="63ca6054-02"

The plan to fix this is to detect and remove the incorrect ZFS member
entry for the whole disk device from the FS_Info cache of blkid data.

GParted probes storage like this:
1.  Gets a list of device names.  Firstly from the the command line,
    otherwise whole disk device names from /proc/partitions using
    Proc_Partitions_Info::get_device_paths().
2.  Expands the list of device names to device and partition names via
    Proc_Partitions_Info::get_device_and_partition_paths_for().
3.  Runs blkid on the list of device and partition names using
    FS_Info::load_cache_for_paths().

This commit just makes the list of device and partition names (passed
from step 2 to step 3 above) structured so that partition names are
associated with their parent device name.

[1] util-linux issue 4078 - blkid reports whole disk as zfs_member if
    the ending partition is a zfs_member
    https://github.com/util-linux/util-linux/issues/4078

Closes #310 - GParted incorrectly showing ZFS occupying the whole disk
              instead of only one partition