Also support reading exFAT usage from exfatprogs <= 1.2.2 (#261)
Exfatprogs 1.2.2 and earlier are included in a number of the latest
distributions.  Therefore support reading exFAT usage from the earlier
dump.exfat style output.

    Distribution                       exfatprogs
    Alpine 3.20                        1.2.2
    Debian 12                          1.2.0
    RHEL / Rocky Linux / AlmaLinux 9   1.2.2
    Ubuntu 24.04 LTS                   1.2.2

On Debian 12, example dump.exfat output:
    # dump.exfat /dev/sdb1 | egrep 'exfatprogs version|Volume Length\(sectors\):|Sector|Free Clusters:'
    exfatprogs version : 1.2.0
    Volume Length(sectors):                 524288
    Sector Size Bits:                       9
    Sector per Cluster bits:                3
    Free Clusters:                          65019

Make this a separate commit so that it can be easily reverted in future
when this is no longer required.  It also results in simpler code as the
old shift bits style figures are converted into new direct multiplier
style figures.  Therefore the final if statement checking all the
figures are set, computing and assigning partition usage remains simple
and unchanged.

Closes #261 - Unable to read the contents of exfat file system with
              exfatprogs >= 1.2.3