Prefix xfs member variables with m_ and initialise POD member
To make it clear they are xfs member variables.  Also add a default
constructor with an initialisation list so POD (Plain Old Data) member
variable m_src_used is never left uninitialised, as it was previously
only assigned in xfs::copy().  m_dest_mount_point is a Glib::ustring
object which is default constructed to a zero length string so doesn't
need adding to the xfs default constructor initialisation list.

* C++ FAQ / Should my constructors use "initialization lists" or
  "assignment"?
  https://isocpp.org/wiki/faq/ctors#init-lists
      "Initialization lists.  In fact, constructors should initialize as
      a rule all member objects in the initialization list."