Increase minimum required version of libparted to 3.2 (!131)
GNU Parted 3.2 was released 2014-07-28 [1][2].  The oldest supported
distributions have these versions of GNU Parted (and libparted):
    Distribution       EOL            parted --version
    Debian 11          2026-Aug       3.4
    RHEL 8             2029-Jul       3.2
    SLES 15            2031-Jul [3]   3.2
    Ubuntu 20.04 LTS   2025-Apr       3.3

Even older no longer supported distributions:
    Distribution       EOL            parted --version
    Debian 10          2024-Jun       3.2
    RHEL / CentOS 7    2024-Jun       3.1
    SLES 12 SP5        2024-Oct       3.1
    Ubuntu 18.04 LTS   2023-Apr       3.2

GParted currently has 4 different sets of conditional code depending on
the version of libparted it is compiled against.  With increasing the
minimum version of libparted to 3.2, the 2 sets of bug workarounds for
older versions of libparted can be removed and the 2 sets of enabling
optional features can always be enabled.

Therefore increase the minimum required version of libparted to 3.2.

[1] GNU Parted 3.2 release announcement
    https://lists.gnu.org/archive/html/info-gnu/2014-07/msg00014.html
[2] NEWS file from GNU Parted 3.2
    http://git.savannah.gnu.org/cgit/parted.git/tree/NEWS?h=v3.2
        "Noteworthy changes in release 3.2 (2014-07-28) [stable]"
[3] Product Support Lifecycle Lifecycle Dates by Product, SUSE Linux
    Enterprise Server 15
    https://www.suse.com/lifecycle/#suse-linux-enterprise-server-15

Closes !131 - Increase minimum required version of libparted to 3.2