Don't run btrfs when not found, while setting mount points (#271)
Repeat the same test as the previous commit.  Have a btrfs file system
mounted, while the btrfs executable is not in $PATH.  GParted reports
this to the terminal:
    # gparted
    GParted 1.6.0-git
    configuration --enable-online-resize
    libparted 3.6
    Failed to execute child process "btrfs" (No such file or directory)

The error is because while GParted is discovering the mount points for
the partition containing the mounted btrfs file system it ends up trying
to run the missing btrfs executable.  The back trace for this is:
    #0 Utils::execute_command("btrfs filesystem show '/dev/sdb1'", ...)
    #1 btrfs::get_cache_entry(path="/dev/sdb1")
    #2 btrfs::get_mount_device(path="/dev/sdb1")
    #3 GParted_Core::set_mountpoints_helper(partition, path="/dev/sdb1")
    #4 GParted_Core::set_mountpoints(partition)
    #5 GParted_Core::set_device_partitions(device, lp_device, lp_disk)
    #6 GParted_Core::set_device_from_disk(device, device_path="/dev/sdb")

This call chain via GParted_Core::set_mountpoints_helper() is not
controlled by any of the file system support capabilities.  So instead
fix this by having the btrfs module remember whether the btrfs
executable is found and only execute it in btrfs::get_cache_entry() when
it is found.

Continuing with the above test, open the Partition Information dialog.
This generates another failed to execute child process btrfs error.  The
backtrace for this is:
    #0 Utils::execute_command("btrfs filesystem show '/dev/sdb1'", ...)
    #1 btrfs::get_cache_entry(path="/dev/sdb1")
    #2 btrfs::get_get_members(path="/dev/sdb1")
    #3 Dialog_Partition_Info::Display_Info()
    #4 Dialog_Partition_Info::Dialog_Partition_Info()
    #5 Win_GParted::activate_info()

The fix in this commit also covers this test case too.

Closes #271 - GParted stuck forever at Searching /dev/sda partitions
              when btrfs-progs not installed