Don't run btrfs when not found, during busy detection (#271)
Repeat test 1 from the first commit "Don't hang printing a UTF-8
exception in a none UTF-8 locale (#271)".  Run GParted with a btrfs file
system on disk but the btrfs executable not in $PATH.  GParted reports
this to the terminal:
    # gparted
    GParted 1.6.0-git
    configuration --enable-online-resize
    libparted 3.6
    Failed to execute child process "btrfs" (No such file or directory)

This happens because busy (mounted) status detection for btrfs was
always set to use the btrfs specific method even when the required btrfs
executable was not found.  (fs.busy was always set to FS::EXTERNAL).

(Btrfs has a specific busy detection method to handle multi-device file
systems which can be mounted via any one of the members in the
multi-device file system, which isn't necessarily the block device
currently being queried).

Fix this by only using the btrfs specific busy detection method when the
btrfs executable is found.  As the busy (mounted) state of a file system
is important in determining which operations can and can't be performed
use the imperfect fallback of GParted's built-in file system mounted
detection method which queries /proc/mounts.  Note that this only
detects when a single device btrfs file system is mounted or when a
multi-device btrfs file system is mounted by the device currently being
queried.  It does not detect when a multi-device btrfs file system is
mounted using a different device to the one currently being queried for
it's busy status.

Closes #271 - GParted hangs while searching /dev/sda partitions when
              btrfs-progs is not installed