Don't run btrfs when not found, reading online usage (#271)
Now mount the btrfs file system, still while the btrfs executable is not
in $PATH.  GParted reports this to the terminal:
    # gparted
    GParted 1.6.0-git
    configuration --enable-online-resize
    libparted 3.6
    Failed to execute child process "btrfs" (No such file or directory)
    Failed to execute child process "btrfs" (No such file or directory)

One of those errors is because in the online (mounted) case the btrfs
specific method for reading file system usage is always called, even
when the btrfs executable is not found.  (fs.online_read was always set
to FS::EXTERNAL).

Fix this by only enabling online (mounted) reading of file system usage
when the btrfs executable is found.

Closes #271 - GParted hangs while searching /dev/sda partitions when
              btrfs-progs is not installed