Move ExecFlags enum and operators to OperationDetail.h (!132)
Fundamentally there are two different types of execute_command() in
GParted:
1. Utils::execute_command()
   Mostly used during device scan in order to read file system
   information from the command's stdout or stderr, but sometimes just
   used to execute a command while making a change.  Expected to be fast
   as GParted just waits for the command to complete.
2. FileSystem::execute_command()
   Used as part of applying an operation when stdout and stderr are
   incrementally captured into operation details as the command runs.
   May be long running.

DMRaid::execute_command() is called when applying an operation and the
output is saved into the operation details.  It actually calls
Utils::execute_command() and manually creates the operation details.  It
really ought to be calling FileSystem::execute_command() but can't
because that's protected in the FileSystem class.  In order to allow the
DMRaid module to use the preferred FileSystem::execute_command() move it
to the OperationDetail class allowing it to be used from anywhere
operation details are used.

Break the change up into a number of smaller incremental commits.  This
commit just moves the ExecFlags enumeration and associated operators
into the OperationDetail header.

Closes !132 - Move FileSystem::execute_command() to OperationDetail
              class